Move GroundControl under pages/academy
